Initial Test for object oriented pacman (OOP)
This is a new method of rendering the game and generating objects using classes. This is new to me and I'm not so keen on my work. It looks really sloppy. In the future possibly clean up the code and make it look nicer and more optimised. But that's what we all say :P